--- app/data.py
@@ -1,15 +1,0 @@
-students = [
-    {
-        "name": "Alice",
-        "id": "1004803",
-    },
-    {
-        "name": "Bob",
-        "id": "1004529"
-    },
-    {
-        "name": "Charlie",
-        "id": "1004910",
-        "gpa": 5.0
-    }
-]
